requirements.txt
gcsfs
pandas-gbq
gspread
beautifulsoup4
pandas
google-auth google-auth-oauthlib google-auth-httplib2
